+++ diesel_cli/tests/tests.rs
@@ -3,6 +3,7 @@
 extern crate diesel;
 extern crate regex;
 extern crate tempdir;
+extern crate chrono;
 
 mod setup;
 mod support;
@@ -10,6 +11,7 @@ mod migration_generate;
 mod migration_redo;
 mod migration_revert;
 mod migration_run;
+mod migration_list;
 mod database_drop;
 mod database_setup;
 mod database_reset;